Handle the time-order requirement of LEAP
Since for the Cuomo data we don't have timing information, that
unfortunately means we cannot run LEAP for it; therefore, we drop the
stages `infer_Cuomo_LEAP` and `plot_Cuomo_LEAP`.

For the Yang data, we decided to average the expression values over
samples per group, as defined in `metadata.csv`, and run LEAP on such
fused data.