Integrate noisyR (preliminary)
This calls noisyR at the beginning of the pipeline to create the
filtered and normalized versions of the dataset, therefore requiring
only the raw variant of the counts matrices.

This implicitly fixes the inconsistent naming between the Yang and Cuomo
datasets, so their variants definitions are consolidated into
`src/analysis/data_variants.csv`.

Since noisyR is currently undergoing a public interface revision, this
integration needs to be updated soon, but this will only affect
`src/denoise.R`; I therefore decided to merge this preliminarily to
avoid merge conflicts after updating the pipeline to DVC v2.0.